pyproject.toml
[tool.poetry]
name = "5_corsework"
version = "0.1.0"
description = ""
authors = ["marinaice181 <anfina1@yandex.ru>"]
readme = "README.md"
packages = [{include = "5_corsework"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"